requirements.txt
fastapi
uvicorn
python-dotenv
requests
playwright
google-genai
bs4
